+++ worker/include/DepLibUring.hpp
@@ -15,13 +15,23 @@ public:
 	/* Struct for the user data field of SQE and CQE. */
 	struct UserData
 	{
-		uint8_t store[1500]{};
+		// Pointer to send buffer.
+		uint8_t* store{ nullptr };
+		// Frame len buffer for TCP.
+		uint8_t frameLen[2] = { 0 };
+		// iovec for TCP, first item for framing, second item for payload.
+		struct iovec iov[2];
+		// Send callback.
 		onSendCallback* cb{ nullptr };
+		// Index in userDatas array.
 		size_t idx{ 0 };
 	};
 
 	/* Number of submission queue entries (SQE). */
 	static constexpr size_t QueueDepth{ 1024 * 4 };
+	static constexpr size_t SendBufferSize{ 1500 };
+
+	using SendBuffer = uint8_t[SendBufferSize];
 
 	static bool IsRuntimeSupported();
 	static void ClassInit();
@@ -29,10 +39,11 @@ public:
 	static flatbuffers::Offset<FBS::LibUring::Dump> FillBuffer(flatbuffers::FlatBufferBuilder& builder);
 	static void StartPollingCQEs();
 	static void StopPollingCQEs();
+	static uint8_t* GetSendBuffer();
 	static bool PrepareSend(
-	  int sockfd, const void* data, size_t len, const struct sockaddr* addr, onSendCallback* cb);
+	  int sockfd, const uint8_t* data, size_t len, const struct sockaddr* addr, onSendCallback* cb);
 	static bool PrepareWrite(
-	  int sockfd, const void* data1, size_t len1, const void* data2, size_t len2, onSendCallback* cb);
+	  int sockfd, const uint8_t* data1, size_t len1, const uint8_t* data2, size_t len2, onSendCallback* cb);
 	static void Submit();
 	static void SetActive();
 	static bool IsActive();
@@ -50,10 +61,16 @@ public:
 		flatbuffers::Offset<FBS::LibUring::Dump> FillBuffer(flatbuffers::FlatBufferBuilder& builder) const;
 		void StartPollingCQEs();
 		void StopPollingCQEs();
+		uint8_t* GetSendBuffer();
 		bool PrepareSend(
-		  int sockfd, const void* data, size_t len, const struct sockaddr* addr, onSendCallback* cb);
+		  int sockfd, const uint8_t* data, size_t len, const struct sockaddr* addr, onSendCallback* cb);
 		bool PrepareWrite(
-		  int sockfd, const void* data1, size_t len1, const void* data2, size_t len2, onSendCallback* cb);
+		  int sockfd,
+		  const uint8_t* data1,
+		  size_t len1,
+		  const uint8_t* data2,
+		  size_t len2,
+		  onSendCallback* cb);
 		void Submit();
 		void SetActive()
 		{
@@ -78,6 +95,10 @@ public:
 
 	private:
 		UserData* GetUserData();
+		bool IsDataInSendBuffers(const uint8_t* data) const
+		{
+			return data >= this->sendBuffers[0] && data <= this->sendBuffers[DepLibUring::QueueDepth - 1];
+		}
 
 	private:
 		// io_uring instance.
@@ -88,10 +109,12 @@ public:
 		uv_poll_t* uvHandle{ nullptr };
 		// Whether we are currently sending RTP over io_uring.
 		bool active{ false };
-		// Pre-allocated UserData entries.
-		UserData userDataBuffer[QueueDepth]{};
+		// Pre-allocated UserData's.
+		UserData userDatas[QueueDepth]{};
 		// Indexes of available UserData entries.
 		std::queue<size_t> availableUserDataEntries;
+		// Pre-allocated SendBuffer's.
+		SendBuffer sendBuffers[QueueDepth];
 		// Submission queue entry process count.
 		uint64_t sqeProcessCount{ 0u };
 		// Submission queue entry miss count.